Add a new modularized bot
Lots of updates here
* Clean up dependencies
* Add pino for logging (and pino-pretty for dev/console logging)
* Add .nvmrc to specify node version
* Multiple .env files, per environment
  TODO Fix this somehow, right now it's unwieldy and requires special handling
    for secret things
* Add a whole slew of new files for running the bot via individual commands
  These are all set up as ES6 module files (mjs)
  * `main` as an entry point to the app, handles options and creates the bot
  * `muse-bot` as the actual bot which defines specific actions and delegates
    to individual modules
  * `muse-google` to handle authenticating and interacting with the Google
    Sheets API
  * `muse-discord` to handle interacting with the Discord.js client

This replicates a tiny bit of functionality from (an old version of) test.js,
but mostly just enough to prove that it works and is able to talk to the
services we need.

Big TODOs left around cleaning up logging, esp for dev vs prod environments,
and between separating config and behavior between those environments,
including leaving space for actually secret stuff.