More flexible input parsing and more configurable output
Now the input file can be either comma- or tab-delimited, and can include a
header row or not.

The output can be configured to be comma- or tab-delimited, and can be sorted
into three separate orders.

To emulate the current seeding behavior, use both the `--output-csv-tabs` flag
and the `--output-order bracket` option.

Sample invokation:
./seed.py --output-csv-tabs --output-order bracket input.csv output.csv