Add support for byes
A little messy, but hacks in support for byes, mostly for 96 brackets

I've included both 48 and 96 test files, pared down from the 64 and 128
respectively. Had some issues with the swapping in particular, but I think
it's all worked out now.

Now I just need to come back in and clean some of this nonsense up.