Super Duper Fun
Adds two options:
- `--badness-duper` for setting how much weight to give to early dupe collisions
- `--drop-dupes-first` for prioritizing dropping submissions from submitters
  with more songs already in

These should work pretty well as is, but `--badness-duper` defaults to 20
(which is the same as artist collisions). It will scale down based on both the
number of dupes someone has and the number of people who submitted a song.

This looked alright to me, but could maybe use tweaking.

The warning output now mentions any submitter/dupe collisions within the first
two rounds.

TODO: Also add some badness for two dupes from the same submitter (but owned
by different people) colliding too early, and warnings to go with it.