Fixes the bug with the output order for dropped submissions
I'm pretty sure I was overthinking it before and my super special generic
implementation was buggy anyway. This replaces all that with just doing the
same thing SORTED output does for both ORIGINAL and BRACKET orderings and
then sorts the output into the original order (which is what they both
expect).

I've also switched from `print` to `logger.info` for all the output
statements (except the CSV writer, which will still write to STDOUT by
default). The output will now automatically be written to a log file
(with a Unix timestamp in the filename) at the same time that it's
written to STDERR (and *not* STDOUT, as before, so that we can easily
distinguish between log messages and the actual CSV output that's
written to STDOUT).

Finally, I've created a directory that the logs are written to and put a
.gitignore file in it so that all the log files won't be committed. Happy
to drop that and just commit log files if we find that useful though (I
figured just keeping a record outside of the terminal was enough and we
could send them back and forth if necessary).